services/gateway/pyproject.toml
[project]
name = "substrate-gateway"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    "pyjwt[crypto]>=2.9.0",
    "structlog>=24.4.0",
    "pydantic-settings>=2.6.0",
    "asyncpg>=0.30.0",
    "sse-starlette>=2.1",
    "substrate-common",
]

[tool.uv.sources]
substrate-common = { workspace = true }

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx",
    "testcontainers[postgres]>=4.8",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
check_untyped_defs = false
warn_unused_ignores = false

[[tool.mypy.overrides]]
module = ["asyncpg.*", "jwt.*", "cryptography.*", "testcontainers.*", "sse_starlette.*"]
ignore_missing_imports = true
